+++ NEWS.md
@@ -1,7 +1,7 @@
 # prediction 0.2.9
 
 * Added `prediction.bigLm()` method for "bigLm" objects from **bigFastlm**. (#1)
-* Added `prediction.biglm()` and `prediction.bigglm()` methods for "biglm" and "bigglm" objects from **biglm**. (#1)
+* Added `prediction.biglm()` and `prediction.bigglm()` methods for "biglm" and "bigglm" objects from **biglm**, including those based by `"ffdf"` from **ff**. (#1)
 
 # prediction 0.2.8